Add search functionality and separate to search.js
