+++ src/js/main.js
@@ -1,7 +1,11 @@
 import { openPost, showTitleBody, showUserEmail } from "./info-modal.js";
+import { getSearchResults, displaySearchResults } from "./search.js";
 
 const searchBtn = document.getElementById('headerSearchBtn');
-searchBtn.addEventListener('click', getSearchResults);
+searchBtn.addEventListener('click', function () {
+  getSearchResults()
+  displaySearchResults();
+});
 
 // Fetch posts from api for posts, return as .json data, and pass to displayPosts function.
 async function getPostData() {
@@ -19,6 +23,8 @@ async function manageData () {
   displayPosts(data)
 }
 
+window.onload = manageData;
+
 
 async function getComments () {
   try {
@@ -31,7 +37,7 @@ async function getComments () {
 }
 
 // Loads the blog posts when site is opened.
-window.onload = manageData;
+
 
 // Shows the blog posts with title and body on the main page
 function displayPosts(data) {
@@ -75,17 +81,4 @@ function displayPosts(data) {
   });
 }
 
-async function getSearchResults () {
-  const searchString = document.getElementById('headerSearch').value;
-  console.log(searchString)
-  const posts = await getPostData();
-  console.log(posts);
-
-  const searchResults = posts.filter((post) => {
-    return post.title.includes(searchString);
-  })
-  console.log(searchResults)
-} 
-
-
-export { getComments };
+export { getPostData, getComments, displayPosts };
